+++ packages/yourloops/test/integration/data/patient-builder.data.ts
@@ -34,9 +34,6 @@ import { type PatientAccount } from '../../../lib/patient/models/patient-profile
 import { type PatientSettings } from '../../../lib/patient/models/patient-settings.model'
 import { Gender } from '../../../lib/auth/models/enums/gender.enum'
 import { UserInviteStatus } from '../../../lib/team/models/enums/user-invite-status.enum'
-import { type ITeamMember } from '../../../lib/team/models/i-team-member.model'
-import { TeamMemberRole } from '../../../lib/team/models/enums/team-member-role.enum'
-import { LanguageCodes } from '../../../lib/auth/models/enums/language-codes.enum'
 import { DiabeticProfile } from '../../../lib/patient/models/patient-diabete-profile'
 import { defaultBgClasses, DiabeticType, Unit } from 'medical-domain'
 
@@ -117,32 +114,5 @@ export const buildPatientMetrics = (params: {
     },
     glycemiaIndicators: defaultGlycemiaIndicators,
     medicalData: params.medicalData || defaultMedicalData
-  }
-}
-
-export const buildTeamMemberFromPatient = (patient: Patient, metrics: PatientMetrics, teamId: string, invitationStatus: UserInviteStatus): ITeamMember => {
-  return {
-    userId: patient.userid,
-    teamId,
-    role: TeamMemberRole.patient,
-    profile: {
-      email: patient.profile.email,
-      firstName: patient.profile.firstName,
-      fullName: patient.profile.fullName,
-      lastName: patient.profile.lastName,
-      patient: { birthday: '1980-01-01T10:44:34+01:00', diagnosisType: 'type1' },
-      privacyPolicy: { acceptanceTimestamp: '2021-05-22', isAccepted: true },
-      termsOfUse: { acceptanceTimestamp: '2021-05-22', isAccepted: true },
-      trainingAck: { acceptanceTimestamp: '2022-10-11', isAccepted: true }
-    },
-    settings: null,
-    preferences: { displayLanguageCode: LanguageCodes.En },
-    invitationStatus,
-    email: patient.profile.email,
-    idVerified: false,
-    unreadMessages: patient.hasSentUnreadMessages ? 1 : 0,
-    alarms: metrics?.monitoringAlerts,
-    glycemiaIndicators: metrics?.glycemiaIndicators,
-    medicalData: metrics?.medicalData
   }
 }